+++ tests/roundedmoney.test.ts
@@ -310,7 +310,7 @@ export default class RoundedMoneyTest {
     @Test("it throws an exception when operand is invalid during multiplication")
     public itThrowsAnExceptionWhenOperandIsInvalidDuringMultiplication(operand: any) {
         const money = new RoundedMoney(1, 2, new Currency("EUR"));
-        const throwFn = () => money.multiply(operand as numeric);
+        const throwFn = () => money.multiply(operand);
         Expect(throwFn).toThrow();
     }
 
@@ -367,7 +367,7 @@ export default class RoundedMoneyTest {
     @Test("it throws an exception when operand is invalid during division")
     public itThrowsAnExceptionWhenOperandIsInvalidDuringDivision(operand: any) {
         const money = new RoundedMoney(1, 2, new Currency("EUR"));
-        const throwFn = () => money.divide(operand as numeric);
+        const throwFn = () => money.divide(operand);
         Expect(throwFn).toThrow();
     }
